+++ Makefile
@@ -12,7 +12,7 @@ SRCS	:= ${addprefix src/, minishell.c \
 		${addprefix environement/, split_env_vars.c fill_env_struct.c env_helpers.c } \
 		${addprefix lexer/, lexer.c tag_in_quotes.c remove_spaces.c} \
 		${addprefix parser/, parser.c parse_redirs.c expand_var.c expand_help.c parse_cmd.c parse_errors.c} \
-		${addprefix exec/, execution.c execute_builtin.c init_exec.c} \
+		${addprefix exec/, execution.c execute_builtin.c init_exec.c exec_fds.c ft_dup.c} \
 		${addprefix builtins/, export.c env.c exit.c pwd.c cd.c echo.c} \
 		${addprefix utils/, ft_error.c  clean_minishell.c ft_strjoin_free.c token_list.c}}
