+++ includes/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: derjavec <derjavec@student.42.fr>          +#+  +:+       +#+        */
+/*   By: deniseerjavec <deniseerjavec@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/08 17:03:55 by skanna            #+#    #+#             */
-/*   Updated: 2024/06/12 11:35:29 by derjavec         ###   ########.fr       */
+/*   Updated: 2024/06/13 13:53:20 by deniseerjav      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,7 +78,9 @@ void		free_env(t_lst_env *env, char *name);
 //exec
 void		execution(t_mini *minishell);
 int			execute_builtin(t_mini *minishell, char *command);
-void      init_fds(t_mini *mini);
+int         init_fds(t_mini *mini);
+int	        ft_dup(t_mini *mini, int i);
+int         fill_fd(t_mini *mini);
 
 //builtin fucntions
 void		echo_command(t_mini *minishell);
